requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
boto3==1.35.0
python-multipart==0.0.12
pillow==11.0.0